requirements.txt
boto3==1.34.54
python-dotenv==1.0.1
pyarrow==15.0.0
pydantic==2.6.3